Refactor tool definitions
Instead of defining tools with `defmethod`s, use `deftool` to provide a
tool registry with information about the tool.

Currently this is functionally exactly the same as the current
implementation. Internally, you can define an `invoke` function that
returns arbitrary data, then `output` and `reactions` functions that
operate on the results and return data to the LLM and the FE,
respectively.

However, this allows us to add other useful data declarations when
defining tools. For example, we can add a `requires-confirmation?`
argument, that will allow us to declaratively state that a tool requires
confirmation.