hacky "fix" to `get-metric-detail`
I'm honestly not super familiar with querying so there might be a better
way to do this, and I'm not totally sure I'm doing the right thing here.

But I noticed that I could filter a metric by columns that weren't
previously appearing in the list of fields for the metric.

My confidence that `breakoutable-columns` is the right thing to use here
is super low though. And I'm not at all sure what we're going to do for
things like native queries??